Moved direction strategies to a seperate folder.
