+++ direction/walking/DirectionStrategy.cpp
