+++ direction/walking/DirectionStrategy.h
