+++ math/GradientModel.cpp
@@ -29,12 +29,12 @@
  **/
 #include "GradientModel.h"
 
+#include "direction/walking/DirectionStrategy.h"
 #include "general/OpenMP.h"
 #include "geometry/SubRoom.h"
 #include "geometry/Wall.h"
 #include "mpi/LCGrid.h"
 #include "pedestrian/Pedestrian.h"
-#include "routing/DirectionStrategy.h"
 
 #include <math.h>
