+++ math/VelocityModel.cpp
@@ -28,12 +28,12 @@
  **/
 #include "VelocityModel.h"
 
+#include "direction/walking/DirectionStrategy.h"
 #include "general/OpenMP.h"
 #include "geometry/SubRoom.h"
 #include "geometry/Wall.h"
 #include "mpi/LCGrid.h"
 #include "pedestrian/Pedestrian.h"
-#include "routing/DirectionStrategy.h"
 
 double xRight = 26.0;
 double xLeft  = 0.0;
